19号执行一下手动旋转建筑

+++ src/main/kotlin/cn/cyanbukkit/speed/utils/Utils.kt
@@ -11,16 +11,16 @@ import java.io.IOException
 /**
  * 模拟进度条 25个▇  按照  分数比 返回  绿色▇（已完成的部分§a） 与红色▇（未完成的部分 颜色§c）
  */
-fun Player.getProgressBar(mess: String, time: Double, maxTime: Double) {
+fun Player.getProgressBar(mess: String, time: Int, maxTime: Int) {
     val percent = time / maxTime
     val progress = 25 * percent
     val green = "§a§l"
     val red = "§c§l"
     val sb = StringBuilder()
-    for (i in 0 until progress.toInt()) {
+    for (i in 0 until progress) {
         sb.append(green).append("|")
     }
-    for (i in 0 until 25 - progress.toInt()) {
+    for (i in 0 until 25 - progress) {
         sb.append(red).append("|")
     }
     Title.actionbar(this, "$mess §7${sb} §f${time}s 结束")
